Refactor to avoid a single huge interface file, fix several warnings
