+++ src/lib.rs
@@ -3,6 +3,8 @@ pub mod interface;
 pub mod db;
 pub mod addressing;
 pub mod syscall_interfaces;
+pub mod testbench;
+pub mod callstack;
 
 extern crate num;
 #[macro_use]
